use HTML5 datetime inputs for competition form
the nice thing about these is that we can require them to be set.
Firefox does not support type="date" and type="time" though so I'll have
to look into jQueryUI or another library to make the experieince
consistent. There still need to be error-catching for improper formats
in FF and validation that the end_date is greater than the start_date.