refactor timezones to use ActiveSupport::TimeZone
my sloppy wrapper around TZInfo was suffering from some daylight savings
bugs, so it's time to replace it with something more mature. This also
replaces the dropdown timezone options with ActiveSupports more
'comprehensive' list.

I still haven't figured out the best syntax for creating TimeWithZone
objects, so I parse the input with DateTime and the apply an offset to
save it as the appropriate UTC time.

I renamed attributes on the Competition model to reflect the data types
that currently return. Since there isn't any real data there are no
migrations.

TimeZone printer is entirely unnecessary now but it felt wrong to have
view-specific formatting in the Competition model, so TimeHelpers were
added to consistent presentation.