persist cookies for 48 hours
I chose 48 hours to encourage competitions to be short. Most users
should not have to log in more than once per competition. Note that this
involves creating a SECRET_TOKEN in env_keys.yml. A simple way to do
this is using the command

`ruby -rsecurerandom -e "puts SecureRandom.hex(32)"`