implement Team#remove_member
I'm not entirely sure whether or not add_member and remove_member should
save the team model. I feel like the controllers should be responsible
for persisting the objects, plus adding a member could happen in the
same action as editing the team so it shouldn't lead to multiple saves.