use joinable? to test when a team member should be added
add_member? used to just use Team#full? to determine whether or not a
user should be added. This would allow duplicate team members which
would then raise an exception. The validation still needs to stand in
case members are added manually but we shouldn't need to raise an
exception if Team#add_member is used properly.