put input for repo url on the team form
a repository url is not mandatory when creating a team (as the team
members might not have a name for their repo until after the competition
has started) but if they have one ready they can add it in.

note: I need to figure out a good way to mark required fields on forms.