update the styles for the main page
I was tired of seeing the main page all cramped together. Since the HTML
of this page is still all just filler, the structure may change in the
future. Right now all of the articles have a similar layout and when I
add a system for posting a reading entries I'll be able to make it
dynamic.