redirect back when logging in
logging in would aways return the user to the home page, but Sinatra has
simple 'redirect back' support which is much nicer. using the same
feature when logging out can run into trouble (if the user logs out on a
page that requires them to be logged in), so that still redirects them
to the home page.