add team members with ajax
since we only need to update the current team members when a new user
joins a team, we can load them in with ajax. the simple solution here is
to load in the partial of the members, although that still leaves the
member order by last updated_at instead of most recently joined.

in order to ensure the 'leave team' button handles the click event the
success callback runs warnIrreversible again. this means that I had to
bring it out of the immediately invoked function it was in so that it
would be visibile in the namespace. the reason I chose not to use jQuery
.on() is because that caused my tests to fail. if you can find a
solution to that problem please let me know!